+++ app/services/forest_liana/resources_getter.rb
@@ -137,15 +137,13 @@ module ForestLiana
         LiveQueryChecker.new(@params[:segmentQuery], 'Live Query Segment').validate()
 
         begin
-          results = ActiveRecord::Base.connection.execute(@params[:segmentQuery])
+          segmentQuery = @params[:segmentQuery].gsub(/\;$/, '')
+          @records = @records.where("#{@resource.table_name}.id IN (#{segmentQuery})")
         rescue => error
           error_message = "Live Query Segment: #{error.message}"
           FOREST_LOGGER.error(error_message)
           raise ForestLiana::Errors::LiveQueryError.new(error_message)
         end
-
-        record_ids = results.to_a.map { |record| record['id'] }
-        @records = @records.where(id: record_ids)
       end
 
       @records = search_query